% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationBenchmarkFunctions.R
\name{benchmarkGeneratorFSP}
\alias{benchmarkGeneratorFSP}
\title{Create Flow shop Scheduling Problem (FSP) Benchmark}
\usage{
benchmarkGeneratorFSP(a, n, m)
}
\arguments{
\item{a}{matrix of processing times for each step and each machine}

\item{n}{number of jobs}

\item{m}{number of machines}
}
\value{
the function of type cost=f(permutation)
}
\description{
Creates a benchmark function for the Flow shop Scheduling Problem.
}
\examples{
n=10
m=4
#ceate a matrix of processing times
A <- matrix(sample(100,replace=TRUE),n,m) 
#create FSP objective function 
fun <- benchmarkGeneratorFSP(A,n,m)
#evaluate
fun(1:n)
fun(n:1)

}
\seealso{
\code{\link{benchmarkGeneratorQAP}}, \code{\link{benchmarkGeneratorTSP}}, \code{\link{benchmarkGeneratorWT}}
}

