% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{combinatorialKriging}
\alias{combinatorialKriging}
\title{Kriging Model}
\usage{
combinatorialKriging(x, y, distanceFunction, control = list())
}
\arguments{
\item{x}{list of samples in input space}

\item{y}{column vector of observations for each sample}

\item{distanceFunction}{a suitable distance function of type f(x1,x2), returning a scalar distance value}

\item{control}{options for the model building procedure}
}
\description{
DEPRECATED version of the Kriging model, please use \code{\link{modelKriging}}
}
\keyword{internal}

