% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRBFN.R
\name{combinatorialRBFN}
\alias{combinatorialRBFN}
\title{Radial Basis Function Network}
\usage{
combinatorialRBFN(x, y, distanceFunction, control = list())
}
\arguments{
\item{x}{list of samples in input space}

\item{y}{column vector of observations for each sample}

\item{distanceFunction}{a suitable distance function of type f(x1,x2), returning a scalar distance value}

\item{control}{options for the model building procedure}
}
\description{
DEPRECATED version of the RBFN model, please use \code{\link{modelRBFN}}
}
\keyword{internal}
