% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indefiniteLearning.R
\name{correctionAugmentedDistanceVector}
\alias{correctionAugmentedDistanceVector}
\title{Augmented Distance Correction}
\usage{
correctionAugmentedDistanceVector(d, object, x)
}
\arguments{
\item{d}{new distance vector}

\item{object}{a modelKriging fit}

\item{x}{new samples (belonging to distances d)}
}
\value{
vector of augmented, corrected distances
}
\description{
Correct new (test) distances, via correcting the augmented distance matrix. Internal use only.
}
\references{
Martin Zaefferer and Thomas Bartz-Beielstein. (2016). Efficient Global Optimization with Indefinite Kernels. Parallel Problem Solving from Nature-PPSN XIV. Accepted, in press. Springer.
}
\keyword{internal}
