% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indefiniteLearning.R
\name{correctionAugmentedKernelVector}
\alias{correctionAugmentedKernelVector}
\title{Augmented Kernel Correction}
\usage{
correctionAugmentedKernelVector(k, object, x)
}
\arguments{
\item{k}{new kernel value vector}

\item{object}{a modelKriging fit}

\item{x}{new samples (belonging to kernel values k)}
}
\value{
vector of augmented, corrected kernel values
}
\description{
Correct new (test) kernel values, via correcting the augmented kernel matrix. Internal use only.
}
\references{
Martin Zaefferer and Thomas Bartz-Beielstein. (2016). Efficient Global Optimization with Indefinite Kernels. Parallel Problem Solving from Nature-PPSN XIV. Accepted, in press. Springer.
}
\keyword{internal}
