% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distanceCalculation.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Calculate Distance Matrix}
\usage{
distanceMatrix(X, distFun)
}
\arguments{
\item{X}{list of samples, where each list element is a suitable input for \code{distFun}}

\item{distFun}{Distance function of type f(x,y)=r, where r is a scalar and x and y are elements whose distance is evaluated.}
}
\value{
The distance matrix
}
\description{
Calculate the distance between all samples in a list, and return as matrix.
}
\examples{
x <- list(5:1,c(2,4,5,1,3),c(5,4,3,1,2), sample(5))
distanceMatrix(x,distancePermutationHamming)
}

