% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutationDistances.r
\name{distancePermutationLCSeq}
\alias{distancePermutationLCSeq}
\title{Longest Common Subsequence Distance for Permutations}
\usage{
distancePermutationLCSeq(x, y)
}
\arguments{
\item{x}{first permutation (integer vector)}

\item{y}{second permutation (integer vector)}
}
\value{
numeric distance value \deqn{d(x,y)}, scaled to values between 0 and 1 (based on the maximum possible distance between two permutations)
}
\description{
DEPRECATED, see \code{\link{distancePermutationInsert}}.
}
\keyword{internal}

