% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceDistances.R
\name{distanceSequenceLevenshtein}
\alias{distanceSequenceLevenshtein}
\title{Levenshtein Distance forsSequences of numbers}
\usage{
distanceSequenceLevenshtein(x, y)
}
\arguments{
\item{x}{first vector (numeric vector)}

\item{y}{second vector (numeric vector)}
}
\value{
numeric distance value \deqn{d(x,y)}
}
\description{
Levenshtein distance for two sequences of numbers
}
\examples{
#e.g., used for distance between integer sequence
x <- c(0,1,10,2,4)
y <- c(10,1,0,4,-4)
distanceSequenceLevenshtein(x,y)
p <- replicate(10,sample(1:5,3,replace=TRUE),simplify=FALSE)
distanceMatrix(p,distanceSequenceLevenshtein)

}
