% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringDistances.R
\name{distanceStringHamming}
\alias{distanceStringHamming}
\title{Hamming Distance for Strings}
\usage{
distanceStringHamming(x, y)
}
\arguments{
\item{x}{first string (class: character)}

\item{y}{second string (class: character)}
}
\value{
numeric distance value \deqn{d(x,y)}
}
\description{
Number of unequal letters in two strings.
}
\examples{
distanceStringHamming("ABCD","AACC")

}
