% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringDistances.R
\name{distanceStringLCStr}
\alias{distanceStringLCStr}
\title{Longest Common Substring distance}
\usage{
distanceStringLCStr(x, y)
}
\arguments{
\item{x}{first string (class: character)}

\item{y}{second string (class: character)}
}
\value{
numeric distance value \deqn{d(x,y)}
}
\description{
Distance between strings, based on the longest common substring.
}
\examples{
distanceStringLCStr("ABCD","AACC")

}
