% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringDistances.R
\name{distanceStringLevenshtein}
\alias{distanceStringLevenshtein}
\title{Levenshtein Distance for Strings}
\usage{
distanceStringLevenshtein(x, y)
}
\arguments{
\item{x}{first string (class: character)}

\item{y}{second string (class: character)}
}
\value{
numeric distance value \deqn{d(x,y)}
}
\description{
Number of insertions, deletions and substitutions to transform one string into another
}
\examples{
distanceStringLevenshtein("ABCD","AACC")

}
