% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrCubic}
\alias{fcorrCubic}
\title{Cubic Kernel for Kriging}
\usage{
fcorrCubic(D, theta = 0)
}
\arguments{
\item{D}{distance matrix}

\item{theta}{kernel parameter}
}
\value{
matrix (Psi)
}
\description{
Cubic Kernel for Kriging
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}
