% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrGauss}
\alias{fcorrGauss}
\title{Gaussian Kernel for Kriging}
\usage{
fcorrGauss(theta, D)
}
\arguments{
\item{theta}{kernel parameter}

\item{D}{distance matrix}
}
\value{
matrix (Psi)
}
\description{
Gaussian Kernel for Kriging
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}

