% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrLinear}
\alias{fcorrLinear}
\title{Linear Kernel for Kriging}
\usage{
fcorrLinear(theta, D)
}
\arguments{
\item{theta}{kernel parameter}

\item{D}{distance matrix}
}
\value{
matrix (Psi)
}
\description{
Linear Kernel for Kriging
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}

