% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrSphere}
\alias{fcorrSphere}
\title{Spherical Kernel for Kriging}
\usage{
fcorrSphere(theta, D)
}
\arguments{
\item{theta}{kernel parameter}

\item{D}{distance matrix}
}
\value{
matrix (Psi)
}
\description{
Spherical Kernel for Kriging
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}

