% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{infillExpectedImprovement}
\alias{infillExpectedImprovement}
\title{Negative Logarithm of Expected Improvement}
\usage{
infillExpectedImprovement(mean, sd, min)
}
\arguments{
\item{mean}{predicted mean values}

\item{sd}{predicted standard deviation}

\item{min}{minimum of all observations so far}
}
\value{
Returns the negative logarithm of the Expected Improvement.
}
\description{
This function calculates the Expected Improvement" of candidate solutions,
based on predicted means, standard deviations (uncertainty) and
the best known objective function value so far.
}
