% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{initializeDesign}
\alias{initializeDesign}
\title{Initialize Random Design}
\usage{
initializeDesign(par = NULL, cF, size)
}
\arguments{
\item{par}{Optional list of user specified solutions to be added to the design/population, defaults to NULL}

\item{size}{size of the design/population}

\item{cf}{Creation function, creates random new individuals}
}
\value{
Returns list with experimental design (or population) without duplicates
}
\description{
Create a random initial population or experimental design, given a specifed creation function,
as well as a optional set of user-specified design members and a maximum design size.
Also removes duplicates from the design/population.
}
\keyword{internal}

