% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceBasedLandscapeGenerators.R
\name{landscapeGeneratorGaussianEval}
\alias{landscapeGeneratorGaussianEval}
\title{Gaussian Landscape Evaluation}
\usage{
landscapeGeneratorGaussianEval(x, glg)
}
\arguments{
\item{x}{list of samples to evaluate}

\item{glg}{list of values defining the Gaussian Landscape, created by \code{landscapeGeneratorGaussianBuild}.}
}
\value{
returns a list, with the following items:\cr
\code{value} value of the combined landscape
\code{components} value of each component
}
\description{
Evaluate a Gaussian landscape. Should not be called directly.
}
\seealso{
\code{\link{landscapeGeneratorGaussian}}
}
\keyword{internal}
