% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationDistances.r
\name{lexicographicPermutationOrderNumber}
\alias{lexicographicPermutationOrderNumber}
\title{Lexicographic order number}
\usage{
lexicographicPermutationOrderNumber(x)
}
\arguments{
\item{x}{permutation (integer vector)}
}
\value{
numeric value giving position in lexicographic order.
}
\description{
This function returns the position-number that a permutation would receive in a lexicographic ordering.
It is used in the lexicographic distance measure.
}
\examples{
lexicographicPermutationOrderNumber(1:5)
lexicographicPermutationOrderNumber(c(1,2,3,5,4))
lexicographicPermutationOrderNumber(c(1,2,4,3,5))
lexicographicPermutationOrderNumber(c(1,2,4,5,3))
lexicographicPermutationOrderNumber(c(1,2,5,3,4))
lexicographicPermutationOrderNumber(c(1,2,5,4,3))
lexicographicPermutationOrderNumber(c(1,3,2,4,5))
lexicographicPermutationOrderNumber(5:1)
lexicographicPermutationOrderNumber(1:7)
lexicographicPermutationOrderNumber(7:1)

}
\seealso{
\code{\link{distancePermutationLex}}
}

