% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutationDistances.r
\name{longestIncreasingSubsequence}
\alias{longestIncreasingSubsequence}
\title{Longest Increasing Subsequence}
\usage{
longestIncreasingSubsequence(x)
}
\arguments{
\item{x}{integer vector}
}
\description{
Calculate the longest increasing subsequence (LIS) in a vector (e.g., a permutation).
}
\examples{
longestIncreasingSubsequence(1:5)
longestIncreasingSubsequence(5:1)
longestIncreasingSubsequence(c(1,2,4,3,5))
longestIncreasingSubsequence(c(1,2,4,3,5,1,2,4,3,2,9))
}
\references{
Wikipedia contributors, "Longest increasing subsequence", Wikipedia, The Free Encyclopedia, 12 November 2014, 19:38 UTC, <http://en.wikipedia.org/w/index.php?title=Longest_increasing_subsequence&oldid=633565014> [accessed 13 November 2014]
}

