% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKriging.R
\name{modelKrigingClust}
\alias{modelKrigingClust}
\title{Build clustered model}
\usage{
modelKrigingClust(x, y, distanceFunction, control = list())
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{distanceFunction}{distanceFunction}

\item{control}{control}
}
\value{
an object
}
\description{
This function builds an ensemble of Gaussian Process model, where
each individual model is fitted to a partition of the parameter space.
Partitions are generated by.
}
