% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKrigingPredict.R
\name{modelKrigingInternalPredictor}
\alias{modelKrigingInternalPredictor}
\title{Kriging Prediction (internal)}
\usage{
modelKrigingInternalPredictor(object, x)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{modelKriging}.}

\item{x}{list of samples to be predicted}

\item{...}{further arguments, not used}
}
\value{
returns a list with:
\describe{
\item{\code{y}}{predicted values}
\item{\code{psi}}{correlations between x and training data}
}
}
\description{
Predict with a model fit resulting from \code{\link{modelKriging}}.
}
\seealso{
\code{\link{simulate.modelKriging}}

\code{\link{predict.modelKriging}}
}
\keyword{internal}
