% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{mutationBinary}
\alias{mutationBinary}
\title{Bit-flip Mutation for Bit-strings}
\usage{
mutationBinary(population, parameters)
}
\arguments{
\item{population}{List of bit-strings}

\item{parameters}{list of parameters: parameters$mutationRate => mutation rate, specifying number of bits flipped. Should be in range between zero and one}
}
\value{
mutated population
}
\description{
Given a population of bit-strings, this function mutates all
individuals by randomly inverting one or more bits in each individual.
}

