% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{mutationBinaryCycle}
\alias{mutationBinaryCycle}
\title{Cycle Mutation for Bit-strings}
\usage{
mutationBinaryCycle(population, parameters)
}
\arguments{
\item{population}{List of bit-strings}

\item{parameters}{list of parameters: parameters$mutationRate => mutation rate, specifying number of bits flipped. Should be in range between zero and one}
}
\value{
mutated population
}
\description{
Given a population of bit-strings, this function mutates all 
individuals by cyclical shifting the string to the right or left.
}
