% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{mutationBinaryFast}
\alias{mutationBinaryFast}
\title{Bit-flip Mutation for Bit-strings (Fast)}
\usage{
mutationBinaryFast(population, parameters)
}
\arguments{
\item{population}{List of bit-strings}

\item{parameters}{not used}
}
\value{
mutated population
}
\description{
Given a population of bit-strings, this function mutates all
individuals by randomly inverting one bit in each individual.
Due to the fixed mutation rate, this is computationally faster.
}

