% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{mutationPermutationInterchangeCore}
\alias{mutationPermutationInterchangeCore}
\title{Interchange of permutation elements}
\usage{
mutationPermutationInterchangeCore(
  population,
  popsize,
  mutations,
  index1,
  index2
)
}
\arguments{
\item{population}{List of permutations}

\item{popsize}{population size}

\item{mutations}{number of mutated elements for each individual}

\item{index1}{vector of first indices, one element for each interchange}

\item{index2}{vector of second indices, one element for each interchange}
}
\value{
mutated population
}
\description{
Support function for \code{\link{mutationPermutationInterchange}} and \code{\link{mutationPermutationSwap}}.
}
\keyword{internal}
