% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{mutationPermutationReversal}
\alias{mutationPermutationReversal}
\title{Reversal Mutation for Permutations}
\usage{
mutationPermutationReversal(population, parameters = list())
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{list of parameters, currently only uses parameters$mutationRate, 
which should be between 0 and 1 (but can be larger than 1). The mutation rate determines the number of reversals
performed, relative to the permutation length (N). 0 means none. 1 means N reversals.
The default is 1/N.}
}
\value{
mutated population
}
\description{
Given a population of permutations, this function mutates all 
individuals by randomly selecting two indices, and reversing the respective sub-permutation.
}

