% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{mutationPermutationSwap}
\alias{mutationPermutationSwap}
\title{Swap Mutation for Permutations}
\usage{
mutationPermutationSwap(population, parameters)
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{not used}
}
\value{
mutated population
}
\description{
Given a population of permutations, this function mutates all
individuals by randomly interchanging two adjacent elements of the permutation.
}

