% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperators.R
\name{mutationStringRandomChange}
\alias{mutationStringRandomChange}
\title{Mutation for Strings}
\usage{
mutationStringRandomChange(population, parameters = list())
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{list of parameters, with \code{parameters$mutationRate} and \code{parameters$lts}.
\code{parameters$mutationRate} should be between 0 and 1 (but can be larger than 1). The mutation rate determines the number of interchanges
performed, relative to the permutation length (N). 0 means none. 1 means N interchanges.
The default is 1/N. \code{parameters$lts} are the possible letters in the string.}
}
\value{
mutated population
}
\description{
Given a population of strings, this function mutates all 
individuals by randomly changing an element of the string.
}
