% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimCEGO.R
\name{optimizeModel}
\alias{optimizeModel}
\title{Optimize Surrogate Model}
\usage{
optimizeModel(res, creationFunction, model, control)
}
\arguments{
\item{res}{TODO}

\item{creationFunction}{TODO}

\item{model}{TODO}
}
\value{
TODO
}
\description{
Interface to the optimization of the surrogate model
}
\seealso{
\code{\link{optimCEGO}}
}
\keyword{internal}

