% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{predict.CKriging}
\alias{predict.CKriging}
\title{Predict: Combinatorial Kriging}
\usage{
\method{predict}{CKriging}(object, x, predAll = FALSE, ...)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{CKriging}.}

\item{x}{list of samples to be predicted}

\item{predAll}{if TRUE return all (RMSE and prediction, in a dataframe), else return only prediction}

\item{...}{further arguments, not used}
}
\value{
Returned value depends on the setting of \code{predAll}\cr
TRUE: data.frame with 2 columns for function value \code{f} and uncertianty estimate \code{s} (RMSE)\cr
FALSE: data.frame with column \code{f} only
}
\description{
Predict with a model fit resulting from \code{\link{combinatorialKriging}}.
}
\seealso{
\code{\link{combinatorialKriging}}
}

