% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelLm.R
\name{predict.CLM}
\alias{predict.CLM}
\title{Predict: Combinatorial Kriging}
\usage{
\method{predict}{CLM}(object, x, ...)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{CLM}.}

\item{x}{list of samples to be predicted}

\item{...}{further arguments, not used}
}
\value{
numeric vector of predictions
}
\description{
Predict with acombinatorialLM fit.
}
\seealso{
\code{\link{combinatorialLM}}
}

