% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterKriging.R
\name{predict.modelKrigingClust}
\alias{predict.modelKrigingClust}
\title{Clustered Kriging Prediction}
\usage{
\method{predict}{modelKrigingClust}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the clustered Kriging model ensemble (settings and parameters), of class \code{modelKrigingClust}.}

\item{newdata}{list of samples to be predicted}

\item{...}{further arguments, currently not used}
}
\value{
list with function value (mean) \code{object$y} and uncertainty estimate \code{object$s} (standard deviation)\cr
}
\description{
Predict with a model fit resulting from \code{\link{modelKrigingClust}}.
}
\seealso{
\code{\link{predict.modelKriging}}
}
