% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelLm.R
\name{predict.modelLinear}
\alias{predict.modelLinear}
\title{Predict: Combinatorial Kriging}
\usage{
\method{predict}{modelLinear}(object, x, ...)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{modelLinear}.}

\item{x}{list of samples to be predicted}

\item{...}{further arguments, not used}
}
\value{
numeric vector of predictions
}
\description{
Predict with amodelLinear fit.
}
\seealso{
\code{\link{modelLinear}}
}

