% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelRBFN.R
\name{predict.modelRBFN}
\alias{predict.modelRBFN}
\title{Predict: Combinatorial RBFN}
\usage{
\method{predict}{modelRBFN}(object, x, ...)
}
\arguments{
\item{object}{fit of the RBFN model (settings and parameters), of class \code{modelRBFN}.}

\item{x}{list of samples to be predicted}

\item{...}{further arguments, not used}
}
\value{
Returned value depends on the setting of \code{object$predAll}\cr
TRUE: list with function value (mean) \code{$y} and uncertainty estimate \code{$s} (standard deviation)\cr
FALSE:\code{$y}only
}
\description{
Predict with a model fit resulting from \code{\link{modelRBFN}}.
}
\seealso{
\code{\link{modelRBFN}}
}

