% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKrigingPrint.R
\name{print.modelKriging}
\alias{print.modelKriging}
\title{Print Function: modelKriging}
\usage{
\method{print}{modelKriging}(x, ...)
}
\arguments{
\item{x}{fit returned by \code{\link{modelKriging}}.}

\item{...}{additional parameters}
}
\description{
Print information about a Kriging fit, as produced by \code{\link{modelKriging}}.
}
\keyword{internal}
