% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{recombinationBinaryAnd}
\alias{recombinationBinaryAnd}
\title{Arithmetic (AND) Crossover for Bit Strings}
\usage{
recombinationBinaryAnd(population, parameters)
}
\arguments{
\item{population}{List of bit-strings}

\item{parameters}{not used}
}
\value{
population of recombined offspring
}
\description{
Given a population of bit-strings, this function recombines each
individual with another individual by computing \code{parent1 & parent2} (logical AND).
}
