% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{recombinationBinaryUniform}
\alias{recombinationBinaryUniform}
\title{Uniform Crossover for Bit Strings}
\usage{
recombinationBinaryUniform(population, parameters)
}
\arguments{
\item{population}{List of bit-strings}

\item{parameters}{not used}
}
\value{
population of recombined offspring
}
\description{
Given a population of bit-strings, this function recombines each
individual with another individual by randomly picking bits from each parent. 
Note, that \code{\link{optimEA}} will not pass the whole population
to recombination functions, but only the chosen parents.
}
