% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{recombinationPermutationAlternatingPosition}
\alias{recombinationPermutationAlternatingPosition}
\title{Alternating Position Crossover (AP) for Permutations}
\usage{
recombinationPermutationAlternatingPosition(population, parameters)
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{not used}
}
\value{
population of recombined offspring
}
\description{
Given a population of permutations, this function recombines each
individual with another individual.
Note, that \code{\link{optimEA}} will not pass the whole population
to recombination functions, but only the chosen parents.
}
