% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{recombinationPermutationCycleCrossover}
\alias{recombinationPermutationCycleCrossover}
\title{Cycle Crossover (CX) for Permutations}
\usage{
recombinationPermutationCycleCrossover(population, parameters)
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{not used}
}
\value{
population of recombined offspring
}
\description{
Given a population of permutations, this function recombines each
individual with another random individual.
Note, that \code{\link{optimEA}} will not pass the whole population
to recombination functions, but only the chosen parents.
}

