% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperators.R
\name{recombinationStringSinglePointCrossover}
\alias{recombinationStringSinglePointCrossover}
\title{Single Point Crossover for Strings}
\usage{
recombinationStringSinglePointCrossover(population, parameters)
}
\arguments{
\item{population}{List of strings}

\item{parameters}{not used}
}
\value{
population of recombined offspring
}
\description{
Given a population of strings, this function recombines each
individual with another random individual.
Note, that \code{\link{optimEA}} will not pass the whole population
to recombination functions, but only the chosen parents.
}
