% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove Duplicates}
\usage{
removeDuplicates(x, cf)
}
\arguments{
\item{x}{List of individuals}

\item{cf}{Creation function, creates random new individuals}
}
\value{
Returns \code{x} without duplicates
}
\description{
Remove duplicates in \code{x}, replace with non-duplicated individuals according to \code{cf}.
}
\keyword{internal}

