% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{removeDuplicatesOffspring}
\alias{removeDuplicatesOffspring}
\title{Remove Duplicates from Offsprings}
\usage{
removeDuplicatesOffspring(xhist, off, cf)
}
\arguments{
\item{xhist}{List of previous individuals}

\item{off}{List of offspring individuals}

\item{cf}{Creation function, creates random new individuals}
}
\value{
Returns \code{off} without duplicates
}
\description{
Remove duplicates in \code{c(xhist,off)}, replace with non-duplicated individuals according to \code{cf}.
}
\keyword{internal}

