% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryOperators.R
\name{solutionFunctionGeneratorBinary}
\alias{solutionFunctionGeneratorBinary}
\title{Binary String Generator Function}
\usage{
solutionFunctionGeneratorBinary(N)
}
\arguments{
\item{N}{length of the bit-strings}
}
\value{
returns a function, without any arguments
}
\description{
Returns a function that generates random bit-strings of length N.
Can be used to create individuals of NK-Landscapes or other problems with binary representation.
}
