% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOperators.R
\name{solutionFunctionGeneratorString}
\alias{solutionFunctionGeneratorString}
\title{String Generator Function}
\usage{
solutionFunctionGeneratorString(N, lts = c("A", "C", "G", "T"))
}
\arguments{
\item{N}{length of the permutations returned}

\item{lts}{letters allowed in the string}
}
\value{
returns a function, without any arguments
}
\description{
Returns a function that generates random strings of length N, with given letters.
Can be used to generate individual solutions for permutation problems, e.g., Travelling Salesperson Problem
}
\examples{
fun <- solutionFunctionGeneratorString(10,c("A","C","G","T"))
fun()
fun()
fun()

}
