% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim2opt.R
\name{step2Opt}
\alias{step2Opt}
\title{2-Opt Step}
\usage{
step2Opt(route, i, k)
}
\arguments{
\item{route}{route to be partially reversed}

\item{i}{start of reversal}

\item{k}{end of reversal}
}
\value{
a new route
}
\description{
Helper function: A single 2-opt step for \code{\link{optim2Opt}}
}
\keyword{internal}

