% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tournamentSelection}
\alias{tournamentSelection}
\title{Tournament Selection}
\usage{
tournamentSelection(
  fitness,
  tournamentSize,
  tournamentProbability,
  selectFromN
)
}
\arguments{
\item{fitness}{Fitness values of individuals}

\item{tournamentSize}{Tournament Size}

\item{tournamentProbability}{Tournament Probability}

\item{selectFromN}{Number of tournament winners}
}
\value{
index of tournament winners
}
\description{
Simple Tournament Selection implementation.
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}
