% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cegoPackage.R
\docType{package}
\name{CEGO-package}
\alias{CEGO}
\alias{CEGO-package}
\title{Combinatorial Efficient Global Optimization in R}
\description{
Combinatorial Efficient Global Optimization
}
\details{
Model building, surrogate model
based optimization and Efficient Global Optimization in combinatorial
or mixed search spaces.

\tabular{ll}{
Package: \tab CEGO\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2015-10-06\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\section{Acknowledgments}{

This work has been partially supported by the Federal Ministry of Education
and Research (BMBF) under the grants CIMO (FKZ 17002X11) and
MCIOP (FKZ 17N0311).
}
\author{
Martin Zaefferer \email{mzaefferer@gmail.com}
}
\references{
Zaefferer, Martin; Stork, Joerg; Friese, Martina; Fischbach, Andreas; Naujoks, Boris; Bartz-Beielstein, Thomas. (2014). Efficient global optimization for combinatorial problems. In Proceedings of the 2014 conference on Genetic and evolutionary computation (GECCO '14). ACM, New York, NY, USA, 871-878. DOI=10.1145/2576768.2598282 http://doi.acm.org/10.1145/2576768.2598282

Zaefferer, Martin; Stork, Joerg; Bartz-Beielstein, Thomas. (2014). Distance Measures for Permutations in Combinatorial Efficient Global Optimization. In Parallel Problem Solving from Nature - PPSN XIII (p. 373-383). Springer International Publishing.
}
\seealso{
Interface of main function: \code{\link{optimCEGO}}
}
\keyword{package}

