% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrCubic}
\alias{fcorrCubic}
\title{Cubic Kernel for Kriging}
\usage{
fcorrCubic(theta, D)
}
\arguments{
\item{theta}{kernel parameter}

\item{D}{distance matrix}
}
\value{
matrix (Psi)
}
\description{
Cubic Kernel for Kriging
}
\seealso{
\code{\link{combinatorialKriging}}
}
\keyword{internal}

