% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKriging.R
\name{fcorrLinear}
\alias{fcorrLinear}
\title{Linear Kernel for Kriging}
\usage{
fcorrLinear(D, theta = 0)
}
\arguments{
\item{D}{distance matrix}

\item{theta}{kernel parameter}
}
\value{
matrix (Psi)
}
\description{
Linear Kernel for Kriging
}
\seealso{
\code{\link{modelKriging}}
}
\keyword{internal}
