% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{mutationPermutationInterchange}
\alias{mutationPermutationInterchange}
\title{Interchange Mutation for Permutations}
\usage{
mutationPermutationInterchange(population, parameters = list())
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{list of parameters, currently only uses parameters$mutationRate,
which should be between 0 and 1 (but can be larger than 1). The mutation rate determines the number of interchanges
performed, relative to the permutation length (N). 0 means none. 1 means N interchanges.
The default is 1/N.}
}
\value{
mutated population
}
\description{
Given a population of permutations, this function mutates all
individuals by randomly interchanging two arbitrary elements of the permutation.
}

