% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{mutationPermutationSwap}
\alias{mutationPermutationSwap}
\title{Swap Mutation for Permutations}
\usage{
mutationPermutationSwap(population, parameters = list())
}
\arguments{
\item{population}{List of permutations}

\item{parameters}{list of parameters, currently only uses parameters$mutationRate, 
which should be between 0 and 1 (but can be larger than 1). The mutation rate determines the number of swaps
performed, relative to the permutation length (N). 0 means none. 1 means N swaps.
The default is 1/N.}
}
\value{
mutated population
}
\description{
Given a population of permutations, this function mutates all 
individuals by randomly interchanging two adjacent elements of the permutation.
}

