% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimCEGO.R
\name{optimizeModel}
\alias{optimizeModel}
\title{Optimize Surrogate Model}
\usage{
optimizeModel(res, creationFunction, model, control)
}
\arguments{
\item{res}{result state of the optimization process}

\item{creationFunction}{Function to create individuals/solutions in search space.}

\item{model}{result of the buildModel function}

\item{control}{list of settings, from optimCEGO}
}
\value{
result list of the optimizer
}
\description{
Interface to the optimization of the surrogate model
}
\seealso{
\code{\link{optimCEGO}}
}
\keyword{internal}
