% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelKrigingPredict.R
\name{predict.modelKriging}
\alias{predict.modelKriging}
\title{Kriging Prediction}
\usage{
\method{predict}{modelKriging}(object, x, ...)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{modelKriging}.}

\item{x}{list of samples to be predicted}

\item{...}{further arguments, not used}
}
\value{
Returned value depends on the setting of \code{object$predAll}\cr
TRUE: list with function value (mean) \code{object$y} and uncertainty estimate \code{object$s} (standard deviation)\cr
FALSE:\code{object$y}only
}
\description{
Predict with a model fit resulting from \code{\link{modelKriging}}.
}
\seealso{
\code{\link{modelKriging}}

\code{\link{simulate.modelKriging}}
}
