% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeDuplicatesOffspring}
\alias{removeDuplicatesOffspring}
\title{Remove Duplicates from Offsprings}
\usage{
removeDuplicatesOffspring(xhist, off, cf, df = duplicated)
}
\arguments{
\item{xhist}{List of previous individuals}

\item{off}{List of offspring individuals}

\item{cf}{Creation function, creates random new individuals}

\item{df}{Dupliate Function. This function determines which elements in a list/population are duplicates. By default, this is the duplicated function from R-base.}
}
\value{
Returns \code{off} without duplicates
}
\description{
Remove duplicates in \code{c(xhist,off)}, replace with non-duplicated individuals according to \code{cf}.
}
\keyword{internal}
