% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutationOperators.R
\name{solutionFunctionGeneratorPermutation}
\alias{solutionFunctionGeneratorPermutation}
\title{Permutation Generator Function}
\usage{
solutionFunctionGeneratorPermutation(N)
}
\arguments{
\item{N}{length of the permutations returned}
}
\value{
returns a function, without any arguments
}
\description{
Returns a function that generates random permutations of length N.
Can be used to generate individual solutions for permutation problems, e.g., Travelling Salesperson Problem
}

